feat: 중복 확인 및 삭제 쿼리 작성
